make QueryColumns purely responsible for computing columns from the given query
