+++ src/platform/packages/shared/kbn-esql-validation-autocomplete/src/autocomplete/autocomplete.ts
@@ -38,6 +38,7 @@ import { QueryColumns, getSourcesHelper } from '../shared/resources_helpers';
 import type { ESQLCallbacks } from '../shared/types';
 import { getCommandContext } from './get_command_context';
 import { mapRecommendedQueriesFromExtensions } from './utils/recommended_queries_helpers';
+import { getQueryForFields } from './get_query_for_fields';
 
 type GetColumnMapFn = () => Promise<Map<string, ESQLColumnData>>;
 
@@ -57,7 +58,7 @@ export async function suggest(
   }
 
   const { getColumnsByType, getColumnMap } = getColumnsByTypeRetriever(
-    root,
+    getQueryForFields(correctedQuery, root),
     innerText,
     resourceRetriever
   );
