+++ src/platform/packages/shared/kbn-esql-validation-autocomplete/src/shared/marker_node_removal.test.ts
@@ -74,10 +74,6 @@ describe('it should remove marker nodes from the AST', () => {
     assertMarkerRemoved(`FROM employees | EVAL total = salary + `);
     assertMarkerRemoved(`FROM employees | EVAL total = salary + bonus, `);
 
-    // After assignment
-    // @TODO reenable when fixed...
-    // assertMarkerRemoved(`FROM employees | EVAL total = `);
-
     // STATS command with binary operator and comma
     assertMarkerRemoved(`FROM employees | STATS avg(salary), `);
 
@@ -87,4 +83,6 @@ describe('it should remove marker nodes from the AST', () => {
     // SORT command with comma
     assertMarkerRemoved(`FROM employees | SORT age, `);
   });
+
+  it.todo('removes marker from right-side of assignment'); // e.g. assertMarkerRemoved(`FROM employees | EVAL total = `);
 });
